refactor: remove useless file from scaffolding
Remove Tab which is a welcome page in project
Move 2 useful files which are required for sdk init to util folder
