--- tabs/src/components/Tab.jsx
@@ -1,8 +1,0 @@
-import React from "react";
-import { Welcome } from "./sample/Welcome";
-
-export default function Tab() {
-  return <div>
-    <Welcome showFunction={ false } />
-  </div>;
-}
