--- tabs/src/components/sample/Deploy.css
@@ -1,5 +1,0 @@
-.deploy.page > img {
-  margin: 0 auto;
-  display: block;
-  width: 100%;
-}
