--- tabs/src/components/sample/Publish.css
@@ -1,5 +1,0 @@
-.publish.page > img {
-  margin: 0 auto;
-  display: block;
-  width: 100%;
-}
